Complete Battle System
You are the hero. You are all that is left. Your village has been taken
over while you were traveling abroad exploring the world. You come back
and must save the village. Set in the fictional village of Zenith and
the surrounding area. Zenith is built in a clearing in a wooded area
surrounded on all sides by mountains, believed to be the home to
magnificent caverns. You can go into some of the houses and into the
cave. As you progress through the cave, you fight off the wizards
henchmen, aka the illuminati puppets. At the end of the game, you must
fight the wizard who is actually the head of the illuminati.